BrainDead-Solution/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
timm>=0.6.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
pydicom>=2.3.0
pillow>=9.0.0
radgraph>=0.1.0
pycocoevalcap>=1.2
matplotlib>=3.5.0
tqdm>=4.64.0
requests>=2.28.0
jupyter>=1.0.0
ipykernel>=6.0.0